fixes #198 Support running the server as docker container
Added a Dockerfile to run the Java based Server and fixed the
docker-compose.yml
Added .dockerignore to speedup the build process because only
pre-built *.jar files will be required in the image. Everything
else will be excluded from build context.
Updated README and described the alternative way to run the server.
Added alternative extension mechanism and link to the extension
documentation.